+++ modules/planning/math/polynomial_xd.h
@@ -21,6 +21,7 @@
 #ifndef MODULES_PLANNING_MATH_POLYNOMIAL_XD_H_
 #define MODULES_PLANNING_MATH_POLYNOMIAL_XD_H_
 
+#include <cinttypes>
 #include <vector>
 
 namespace apollo {
@@ -29,15 +30,15 @@ namespace planning {
 class PolynomialXd {
  public:
   PolynomialXd() = default;
-  explicit PolynomialXd(const std::size_t order);
+  explicit PolynomialXd(const std::uint32_t order);
   explicit PolynomialXd(const std::vector<double>& params);
   double operator()(const double value) const;
-  double operator[](const std::size_t index) const;
+  double operator[](const std::uint32_t index) const;
   void set_params(const std::vector<double>& params);
   void derived_from(const PolynomialXd& base);
   void integrated_from(const PolynomialXd& base);
   void integrated_from(const PolynomialXd& base, const double intercept);
-  std::size_t order() const;
+  std::uint32_t order() const;
   const std::vector<double>& params() const;
 
  private:
