+++ modules/planning/math/smoothing_spline/spline_1d.cc
@@ -28,10 +28,11 @@
 namespace apollo {
 namespace planning {
 
-Spline1d::Spline1d(const std::vector<double>& x_knots, const std::size_t order)
+Spline1d::Spline1d(const std::vector<double>& x_knots,
+                   const std::uint32_t order)
     : x_knots_(x_knots), spline_order_(order) {
   if (x_knots.size() > 1) {
-    for (std::size_t i = 1; i < x_knots_.size(); ++i) {
+    for (std::uint32_t i = 1; i < x_knots_.size(); ++i) {
       splines_.emplace_back(spline_order_);
     }
   }
@@ -41,7 +42,7 @@ double Spline1d::operator()(const double x) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(x);
+  std::uint32_t index = find_index(x);
   return splines_[index](x - x_knots_[index]);
 }
 
@@ -50,7 +51,7 @@ double Spline1d::derivative(const double x) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(x);
+  std::uint32_t index = find_index(x);
   return splines_[index].derivative(x - x_knots_[index]);
 }
 
@@ -58,7 +59,7 @@ double Spline1d::second_order_derivative(const double x) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(x);
+  std::uint32_t index = find_index(x);
   return splines_[index].second_order_derivative(x - x_knots_[index]);
 }
 
@@ -66,20 +67,20 @@ double Spline1d::third_order_derivative(const double x) const {
   if (splines_.size() == 0) {
     return 0.0;
   }
-  std::size_t index = find_index(x);
+  std::uint32_t index = find_index(x);
   return splines_[index].third_order_derivative(x - x_knots_[index]);
 }
 
 bool Spline1d::set_spline_segs(const Eigen::MatrixXd& params,
-                               const std::size_t order) {
+                               const std::uint32_t order) {
   // check if the parameter size fit
   if (x_knots_.size() * order !=
-      order + static_cast<std::size_t>(params.rows())) {
+      order + static_cast<std::uint32_t>(params.rows())) {
     return false;
   }
-  for (std::size_t i = 0; i < splines_.size(); ++i) {
+  for (std::uint32_t i = 0; i < splines_.size(); ++i) {
     std::vector<double> spline_piece(order, 0.0);
-    for (std::size_t j = 0; j < order; ++j) {
+    for (std::uint32_t j = 0; j < order; ++j) {
       spline_piece[j] = params(i * order + j, 0);
     }
     splines_[i].set_params(spline_piece);
@@ -88,7 +89,7 @@ bool Spline1d::set_spline_segs(const Eigen::MatrixXd& params,
   return true;
 }
 
-Spline1dSeg* Spline1d::mutable_smoothing_spline(const std::size_t index) {
+Spline1dSeg* Spline1d::mutable_smoothing_spline(const std::uint32_t index) {
   if (index >= splines_.size()) {
     return nullptr;
   } else {
@@ -98,12 +99,12 @@ Spline1dSeg* Spline1d::mutable_smoothing_spline(const std::size_t index) {
 
 const std::vector<double>& Spline1d::x_knots() const { return x_knots_; }
 
-std::size_t Spline1d::spline_order() const { return spline_order_; }
+std::uint32_t Spline1d::spline_order() const { return spline_order_; }
 
-std::size_t Spline1d::find_index(const double x) const {
+std::uint32_t Spline1d::find_index(const double x) const {
   auto upper_bound = std::upper_bound(x_knots_.begin() + 1, x_knots_.end(), x);
-  return std::min(x_knots_.size() - 1,
-                  static_cast<std::size_t>(upper_bound - x_knots_.begin())) -
+  return std::min(static_cast<std::uint32_t>(x_knots_.size() - 1),
+                  static_cast<std::uint32_t>(upper_bound - x_knots_.begin())) -
          1;
 }
